Split computer use backend tests
